Move ClockRateManager to clock package
