+++ common/src/main/java/xyz/eclipseisoffline/customtimecycle/mixin/MinecraftServerMixin.java
@@ -18,7 +18,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import xyz.eclipseisoffline.customtimecycle.ClockRateManager;
+import xyz.eclipseisoffline.customtimecycle.clock.ClockRateManager;
 import xyz.eclipseisoffline.customtimecycle.CustomTimeCycle;
 import xyz.eclipseisoffline.customtimecycle.TimeCycleConfiguration;
 import xyz.eclipseisoffline.customtimecycle.clock.CustomTimeCycleClockMarkers;
