+++ common/src/main/java/xyz/eclipseisoffline/customtimecycle/mixin/ServerClockManagerMixin.java
@@ -19,7 +19,7 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import xyz.eclipseisoffline.customtimecycle.ClockRateManager;
+import xyz.eclipseisoffline.customtimecycle.clock.ClockRateManager;
 import xyz.eclipseisoffline.customtimecycle.clock.ClockInstanceUtil;
 import xyz.eclipseisoffline.customtimecycle.clock.ServerClockManagerUtil;
